feat(UI): UI improvements on all views

+++ src/app/property-management/owner-list/owner-list-item/owner-list-item.component.ts
@@ -1,4 +1,7 @@
-import { Component, EventEmitter, Input, OnInit, Output, Renderer2 } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output, Renderer2, TemplateRef, ViewChild } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { TranslateService } from '@ngx-translate/core';
 import { RolesService } from 'src/app/shared/roles.service';
 import { Owner, Property } from '../../property-management.data';
 import { OwnerListItemService } from './owner-list-item.service';
@@ -15,17 +18,24 @@ export class OwnerListItemComponent implements OnInit {
     properties: Property[] = [];
 
     canRemoveOwner: boolean = false;
+    canEditOwner: boolean = false;
+
+    @ViewChild('deleteConfirmation') confirmationDialogTemplate!: TemplateRef<string>;
 
     constructor(
         private ownerListItem: OwnerListItemService,
         private roles: RolesService,
-        private renderer: Renderer2
+        private renderer: Renderer2,
+        private dialog: MatDialog,
+        private snackbar: MatSnackBar,
+        private translate: TranslateService
     ) { }
 
     ngOnInit() {
         this.roles.roles$.subscribe(roles => {
             if (roles.includes('customer-service')) {
                 this.canRemoveOwner = true;
+                this.canEditOwner = true;
             }
         })
     }
@@ -39,15 +49,44 @@ export class OwnerListItemComponent implements OnInit {
     }
 
     async deleteOwner(owner: Owner) {
-        await this.ownerListItem.deleteOwner(owner);
-        this.ownerDeleted.emit();
+        this.dialog.open(this.confirmationDialogTemplate, {
+            height: '20%',
+            width: '100%'
+        }).afterClosed().subscribe(async (toDelete: boolean) => {
+            if (toDelete) {
+                await this.ownerListItem.deleteOwner(owner);
+                this.ownerDeleted.emit();
+
+                this.snackbar.open(
+                    this.translate.instant('owner_list_item.delete_msg'),
+                    this.translate.instant('owner_list_item.dismiss_msg'),
+                    {
+                        duration: 3000
+                    }
+                );
+            }
+        });
+
+
+    }
+
+    async editOwner(owner: Owner) {
+
     }
 
     showDeleteBtn(deleteBtn: HTMLDivElement) {
         this.renderer.removeStyle(deleteBtn, 'display');
     }
 
+    showEditBtn(editBtn: HTMLDivElement) {
+        this.renderer.removeStyle(editBtn, 'display');
+    }
+
     hideDeleteBtn(deleteBtn: HTMLDivElement) {
         this.renderer.setStyle(deleteBtn, 'display', 'none');
+    }
+
+    hideEditBtn(editBtn: HTMLDivElement) {
+        this.renderer.setStyle(editBtn, 'display', 'none');
     }
 }
